% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{EndSentence}
\alias{EndSentence}
\title{End Sentence}
\usage{
EndSentence(string)
}
\arguments{
\item{string}{Input string}
}
\value{
`string`, punctuated with a final full stop (period).`
}
\description{
End Sentence
}
\examples{
{
EndSentence("Hello World")
# "Hello World."
}

}
\author{
Martin R. Smith
}
