% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{GitLink}
\alias{GitLink}
\alias{GitHubPages}
\title{GitHub Link}
\usage{
GitLink(path, alt = NULL, raw = TRUE, user = getOption("GitHubUser"),
  repo = getOption("GitHubRepo"))

GitHubPages(path = "", alt = NULL, user = getOption("GitHubUser"),
  repo = getOption("GitHubRepo"))
}
\arguments{
\item{path}{Character string specifying path to file in specified GitHub repository}

\item{alt}{Character string specifying link text to display if
knitr output is in HTML format.}

\item{raw}{Logical specifying whether to link to the raw file at
raw.githubusercontent.com.
If `FALSE`, links to the file at github.com.}

\item{user, repo}{Character string specifying GitHub username and repository,
overriding any set in options.  These values need not be provided if 
GitHub details have been set globally using
`options(GitHubUser='ms609', GitHubRepo='MorphoBank')`.}
}
\value{
A link to a file hosted on GitHub, formatted appropriately for HTML
or Latex output
}
\description{
Generates a link to a file on a specified GitHub repository.
}
\section{Functions}{
\itemize{
\item \code{GitHubPages}: A link to `user`.github.io/`repo`/
}}

\examples{
{
  GitLink('ms609', 'MorphoBank')
  options(GitHubUser="ms609", GitHubRepo="MorphoBank")
  GitLink()
}
}
\author{
Martin R. Smith
}
