% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{IsTransformational}
\alias{IsTransformational}
\title{Is Character Transformational}
\usage{
IsTransformational(states)
}
\arguments{
\item{states}{TODO}
}
\value{
A logical specifying whether each character is transformational
}
\description{
Is Character Transformational
}
