% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{MorphoBankExports}
\alias{MorphoBankExports}
\title{MorphoBank exports}
\usage{
MorphoBankExports(path = ".", ...)
}
\arguments{
\item{path}{Path to search}

\item{\dots}{Additional parameters to [list.files]}
}
\value{
A character vector listing the full filenames of all Nexus files
in the specified directory beginning "mbank_", and thus presumed exported
from MorphoBank.
}
\description{
MorphoBank exports
}
\author{
Martin R. Smith
}
