% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{MostRecentNexus}
\alias{MostRecentNexus}
\title{Most recent Nexus file}
\usage{
MostRecentNexus(filenames = MorphoBankExports())
}
\arguments{
\item{filenams}{Character vector listing locations of nexus files exported from MorphoBank}
}
\value{
character of length one, specifying the most recent of the files provided
}
\description{
Reports which of a list of nexus files was exported from MorphoBank most recently
}
\author{
Martin R. Smith
}
