% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{PrintCharacterHeading}
\alias{PrintCharacterHeading}
\title{Print character heading}
\usage{
PrintCharacterHeading(char, charNo, prevElements = "*NONE*",
  thisElements = strsplit(char, ": ")[[1]], nextElements = "*NONE*",
  forceDisplay = character(0), Format = I)
}
\arguments{
\item{char}{Character string specifying the definition of the character}

\item{charNo}{Character string identifying the character's number}

\item{prevElements, thisElements, nextElements}{Character description,
decomposed into its constituent hierarchichal elements}

\item{forceDisplay}{Character vector specifying character descriptions to always
print as a header when an exact match is encountered.}

\item{Format}{Function that is passed a character string, and should return the
string with appropriate formatting (e.g. italicization of genus names).}
}
\value{
A heading for the character in markdown format
}
\description{
Prints a character heading using markdown syntax at an appropriate level in
the hierarchy.
}
\examples{
{
  PrintCharacterHeading("Tail: Colour", 1, "NONE",
  c("Tail", "Colour"), c("Tail", "Colour", "Patterning"))
}


}
\author{
Martin R. Smith
}
