% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{PrintNaughtyInapplicables}
\alias{PrintNaughtyInapplicables}
\title{Print Naughty Inapplicables}
\usage{
PrintNaughtyInapplicables(states)
}
\arguments{
\item{states}{A character vector stating the labels for each character state,
as given by `attr(ReadCharacters(filename), 'state.labels')`}
}
\value{
A warning message for inclusion in a markdown documnet.
}
\description{
Prints a warning message when an inapplicable token is found in a neomorphic character.
}
\author{
Martin R. Smith
}
