% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{PrintStates}
\alias{PrintStates}
\title{Print States}
\usage{
PrintStates(states)
}
\arguments{
\item{states}{A character vector stating the labels for each character state,
as given by `attr(ReadCharacters(filename), 'state.labels')`}
}
\value{
Text listing the states, ready for inclusion in a markdown source document
}
\description{
Prints a summary of the character states for a given character
}
\author{
Martin R. Smith
}
