% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{ReadNotes}
\alias{ReadNotes}
\title{Read Notes}
\usage{
ReadNotes(filename)
}
\arguments{
\item{filename}{A string listing the path to a file exported by MorphoBank}
}
\value{
TBC
}
\description{
Read Notes
}
\author{
Martin R. Smith
}
