% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characters.R
\name{IsTransformational}
\alias{IsTransformational}
\alias{IsNeomorphic}
\title{Is Character Transformational}
\usage{
IsTransformational(states)

IsNeomorphic(states)
}
\arguments{
\item{states}{TODO}
}
\value{
A logical specifying whether each character is transformational
}
\description{
Is Character Transformational
}
\section{Functions}{
\itemize{
\item \code{IsNeomorphic}: The inverse of IsTransformational
}}

