% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{NexusTime}
\alias{NexusTime}
\title{Nexus Time}
\usage{
NexusTime(filename, format = "double")
}
\arguments{
\item{filename}{String specifying path to the file}

\item{format}{Format of output; see below}
}
\value{
The time that `filename` was exported from MorphoBank, according
to its internal comment, either by specifying the year, month, day and time as a double
(`format = 'double'`) or as a string in the format `YYYY-MM-DD hh:mm:ss` (otherwise)
}
\description{
Report time that a Nexus file was exported from MorphoBank
}
\author{
Martin R. Smith
}
